# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Social Media Automation API",
    "version": "1.0.0",
    "description": "Complete API for WhatsApp Business, Facebook Graph, Instagram Business, and Meta Pixel integration",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "code": {
                "type": "string"
              }
            }
          }
        }
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      }
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/whatsapp/send-message": {
      "post": {
        "tags": ["WhatsApp"],
        "summary": "Send WhatsApp message",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phoneNumber": {
                    "type": "string"
                  },
                  "message": {
                    "type": "string"
                  },
                  "sector": {
                    "type": "string",
                    "enum": ["education", "hospitality", "investment"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message sent successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/facebook/create-post": {
      "post": {
        "tags": ["Facebook"],
        "summary": "Create Facebook post",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  },
                  "sector": {
                    "type": "string",
                    "enum": ["education", "hospitality", "investment"]
                  },
                  "link": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post created successfully"
          }
        }
      }
    },
    "/api/instagram/create-post": {
      "post": {
        "tags": ["Instagram"],
        "summary": "Create Instagram post",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "caption": {
                    "type": "string"
                  },
                  "imageUrl": {
                    "type": "string"
                  },
                  "sector": {
                    "type": "string",
                    "enum": ["education", "hospitality", "investment"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post created successfully"
          }
        }
      }
    },
    "/api/meta-pixel/track-event": {
      "post": {
        "tags": ["Meta Pixel"],
        "summary": "Track Meta Pixel event",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "eventName": {
                    "type": "string"
                  },
                  "userData": {
                    "type": "object"
                  },
                  "sector": {
                    "type": "string",
                    "enum": ["education", "hospitality", "investment"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event tracked successfully"
          }
        }
      }
    },
    "/api/analytics/dashboard": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get analytics dashboard",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sector",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["education", "hospitality", "investment"]
            }
          },
          {
            "name": "period",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["day", "week", "month"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard data retrieved successfully"
          }
        }
      }
    }
  }
}